+++ docs/conf.py
@@ -404,4 +404,5 @@ intersphinx_mapping = {
     'boto3': ('https://boto3.amazonaws.com/v1/documentation/api/latest/', None),
     's3fs': ('https://s3fs.readthedocs.io/en/latest/', None),
     'kerchunk': ('https://fsspec.github.io/kerchunk/', None),
+    'numcodecs': ('https://numcodecs.readthedocs.io/en/stable/', None),
 }